New pin and bank mapping, separate include files for midi notes and instrument patches, autodetect switch inversion.
In the Aug/Sept 2022 update new PCBs were developed for lever and cherry key accordions. The key mapping has changed: instead of 6 banks we now have 8 banks.
A new project was created: "LeverHarmonica_EightBanks_Sept2022" so that the old project (LeverHarmonica) is still compatible with our 1st iteration (wooden harmonica). In the new project separate header files are used for the midi note definitions and the instrument patch properties. These files could later be moved to a separate "include" directory so that these can be included by multiple Arduino projects (if needed).